add mac address

+++ src/protocols/ArpParser.cpp
@@ -5,7 +5,7 @@
 #include <cstring>
 #include <ctime>
 #include <utility> // for std::pair
-#include <tuple>   // --- tuple 헤더 추가 ---
+#include <tuple>   // for std::tuple
 
 // Helper Function to format timestamp to ISO 8601 (Cross-platform)
 static std::string format_timestamp_arp(const struct timeval& ts) {
@@ -39,8 +39,8 @@ std::string ArpParser::mac_to_string(const uint8_t* mac) {
     return ss.str();
 }
 
-// --- 수정: 반환 타입을 std::tuple로 변경 (timestamp, json_details, op_code) ---
-std::tuple<std::string, std::string, int> ArpParser::parse(const struct pcap_pkthdr* header, const u_char* arp_payload, int size) {
+// --- 수정: 반환 타입을 std::tuple로 변경 ---
+std::tuple<std::string, std::string, uint16_t> ArpParser::parse(const struct pcap_pkthdr* header, const u_char* arp_payload, int size) {
     if (size < sizeof(ARPHeader)) {
         return {"", "", 0}; // Return empty tuple on failure
     }
@@ -53,7 +53,7 @@ std::tuple<std::string, std::string, int> ArpParser::parse(const struct pcap_pkt
     inet_ntop(AF_INET, (void*)arp_header->spa, spa_str, INET_ADDRSTRLEN);
     inet_ntop(AF_INET, (void*)arp_header->tpa, tpa_str, INET_ADDRSTRLEN);
 
-    uint16_t op_code = ntohs(arp_header->oper); // --- op_code 추출 ---
+    uint16_t op_code = ntohs(arp_header->oper);
 
     details_ss << "{\"op\":" << op_code
                << ",\"smac\":\"" << mac_to_string(arp_header->sha) << "\""
@@ -63,6 +63,6 @@ std::tuple<std::string, std::string, int> ArpParser::parse(const struct pcap_pkt
 
     std::string timestamp_str = format_timestamp_arp(header->ts);
     
-    // --- op_code를 함께 반환 ---
+    // --- 수정: op_code를 튜플에 포함하여 반환 ---
     return {timestamp_str, details_ss.str(), op_code};
 }
